sherpa_backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.15.0
pydantic>=1.9.0,<2.0.0
google-generativeai>=0.3.0
python-dotenv==1.0.0 
gunicorn
requests==2.31.0

elevenlabs
google-cloud-aiplatform
